charbuf impl

+++ build.sh
@@ -4,11 +4,13 @@ BINARY_DIR="${BINARY_DIR:-build/release}"
 GENERATOR="${GENERATOR:-Ninja}"
 
 set -euxo pipefail
-rm -rf "$BINARY_DIR"
-mkdir -p "$BINARY_DIR"
-cmake -S . -B "$BINARY_DIR" -G "$GENERATOR" \
-    --toolchain "$VCPKG_ROOT/scripts/buildsystems/vcpkg.cmake" \
-    -DCMAKE_BUILD_TYPE=Release \
-    -DBUILD_TESTING=ON \
-    -DBUILD_DEMOS=ON 
+if [[ ! -d $BINARY_DIR ]]; then
+    rm -rf "$BINARY_DIR"
+    mkdir -p "$BINARY_DIR"
+    cmake -S . -B "$BINARY_DIR" -G "$GENERATOR" \
+        --toolchain "$VCPKG_ROOT/scripts/buildsystems/vcpkg.cmake" \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DBUILD_TESTING=ON \
+        -DBUILD_DEMOS=ON 
+fi
 cmake --build "$BINARY_DIR"
